refactor: add ParseRobots and ParseSiteMap directly into crawler
